backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "floalzaix",email = "floalzfencing@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.10"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "numpy (>=2.3.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
